util: fix typing

+++ caligo/util/file.py
@@ -4,9 +4,9 @@ from mimetypes import guess_type
 from os.path import join
 from pathlib import Path
 from urllib import parse
-from typing import Optional, Tuple, Union
+from typing import Any, Optional, Tuple
 
-from async_property import async_property
+from pyrogram.types import Message
 
 from .async_helpers import run_sync
 from .misc import human_readable_bytes as human
@@ -14,6 +14,13 @@ from .time import format_duration_td as time, sec
 
 
 class File:
+
+    _content: Any
+    _index_link: Optional[str]
+    _invoker: Any
+    _name: str
+    _path: Path
+    _start_time: Any
 
     def __init__(self, path: Path) -> None:
         self._path = path
@@ -46,47 +53,48 @@ class File:
         return self.path.parent.absolute()
 
     @property
-    def mime_type(self) -> str:
+    def mime_type(self) -> Optional[str]:
         return guess_type(self.path)[0]
 
     @property
-    def content(self) -> None:
+    def content(self) -> Any:
         return self._content
 
     @content.setter
-    def content(self, val):
+    def content(self, val) -> None:
         self._content = val
 
     @property
-    def invoker(self) -> None:
+    def invoker(self) -> Message:
         return self._invoker
 
     @invoker.setter
-    def invoker(self, val):
+    def invoker(self, val) -> None:
         self._invoker = val
 
     @property
-    def index_link(self) -> None:
+    def index_link(self) -> Optional[str]:
+        link = None
         if self._index_link is not None:
             link = join(self._index_link, parse.quote(self.name))
 
         return self._index_link if self._index_link is None else link
 
     @index_link.setter
-    def index_link(self, val):
+    def index_link(self, val) -> None:
         self._index_link = val
 
     @property
-    def start_time(self) -> None:
+    def start_time(self) -> int:
         return self._start_time
 
     @start_time.setter
-    def start_time(self, val):
+    def start_time(self, val) -> None:
         self._start_time = val
 
-    @async_property
-    async def progress_string(self) -> Tuple[Union[str, None], bool, None]:
+    async def progress_string(self) -> Tuple[Optional[str], bool, Optional[str]]:
         file = self.content
+        progress = None
         status, response = await run_sync(file.next_chunk, num_retries=5)
         if status:
             after = sec() - self.start_time
@@ -107,7 +115,7 @@ class File:
                 f"__{human(current)} of {human(size)} @ "
                 f"{human(speed, postfix='/s')}\neta - {time(eta)}__\n\n")
 
-        if response is None:
+        if response is None and progress is not None:
             return progress, False, None
 
         size = response.get("size")
@@ -124,8 +132,9 @@ class File:
 
         done = False
         last_update_time = None
+        link = None
         while not done:
-            progress, done, link = await self.progress_string
+            progress, done, link = await self.progress_string()
             now = datetime.now()
             if invoker is not None and progress is not None and (
                     last_update_time is None or (now - last_update_time
@@ -136,6 +145,6 @@ class File:
 
             await asyncio.sleep(0.1)
 
-        if invoker is not None and update is True:
+        if invoker is not None and update is True and link is not None:
             await invoker.reply(link)
             await invoker.delete()
